restructure cezo_fl files
