+++ cezo_fl/server.py
@@ -7,7 +7,7 @@ from typing import Any, Callable, Iterable, Sequence
 import torch
 
 from cezo_fl.client import AbstractClient
-from cezo_fl.random_gradient_estimator import RandomGradientEstimator
+from cezo_fl.gradient_estimators.random_gradient_estimator import RandomGradientEstimator
 from cezo_fl.run_client_jobs import execute_sampled_clients
 from cezo_fl.typing import CriterionType
 from cezo_fl.util.metrics import Metric
